Add email field to the workspace member details view
